requirements.txt
requests
# Django
django==2.2.10
django-shop==1.1.1
django-bootstrap4>=1.1.1
django-oscar==2.0
sorl-thumbnail==12.6.3
djangorestframework==3.11.0
